[feat]: #43 ボールのロジックを分離し、CPU対戦モードを改善
- ball.jsファイルを新規作成し、ボールの初期化、描画、移動ロジックを移動
- script.jsからボール関連の関数を削除し、ball.jsに移動
- ボールの速度と位置の初期化ロジックを追加
- パドルとボールの衝突判定を追加
- プレイヤーとCPUのスコアを管理する関数を追加
- ゲームのリセットと終了ロジックを追加

この変更により、コードの可読性が向上し、CPU対戦モードが改善されました。
